DTSPB 5505: Move GS scripts into pr_specific folder
